cleanup of read models and processors
