--- Source/Analytics/Read/HealthRisks/HealthRisk.cs
@@ -1,13 +1,0 @@
-using Concepts;
-using Concepts.HealthRisks;
-using Dolittle.ReadModels;
-using System;
-
-namespace Read.HealthRisks
-{
-    public class HealthRisk : IReadModel
-    {
-        public HealthRiskId Id { get; set; }
-        public HealthRiskName Name { get; set; }
-    }
-}
